added near-final implementation of objectinfo modals
there remains a persistent bug with parsing base64 images because we do
a dumb NaN -> null replace in the string content of the checkplot.
this is likely a bad idea. Probably need to go back to ugly dict
walking and checking for NaNs there for checkplotserver's standalone
mode.

one way to get around this might be to dump the JSON for the base64
images separately, dump the rest with agressive NaN -> null
replacement, then concatenate the two (in the correct manner so it
becomes a single JSON)