cachestr for query name/spec for later caching if we end up doing that
This is not as straightforward as it should be. Each user has their own
restrictions on objects and datasets based on their access privileges,
so I can't see a good way that doesn't involve recreating a large
portion of the full query workflow for each cached query. We'd need to
regenerate the object list, regenerate the pickle, regenerate the LC
zip, as well as all the paged pickles, etc.

May revisit in the future.

For now, the checklist is:

- [x] hash based on the deep-sorted JSON repr of the input query
- [ ] figure out how to handle duplicate datasets: scrub prev owner and
  ds info, copy over the pickle, async generate new pages as necessary,
  handle duplicate LC zips as done currently
- [ ] figure out how to handle disallowed objects/datasets as well